OS-6754. Internal server error on flavors api
